refactor(pylon): carve out wire DTO modules
Move pylon request and response wire shapes into tagged DTO modules where the original files mixed handler behavior with serde-only types. Keep the existing handler exports stable while giving the TOPOLOGY DTO carve-out a narrow file boundary.

Kanon-Issue: forkwright/kanon#541

Gate-Passed: cargo fmt --check

Gate-Passed: cargo check -p pylon --all-targets

Gate-Passed: cargo test -p pylon

Gate-Passed: cargo clippy -p pylon --all-targets --no-deps -- -D warnings
