+++ crates/pylon/src/event_bus_dto.rs
@@ -1,0 +1,16 @@
+// WHY: wire DTO
+//! Domain event wire shapes.
+
+use serde::{Deserialize, Serialize};
+
+/// A domain event with a stable topic name and JSON payload.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[non_exhaustive]
+pub struct DomainEvent {
+    /// Event topic (e.g. `fact.created`, `turn.complete`).
+    pub topic: String,
+    /// Structured event payload.
+    pub payload: serde_json::Value,
+    /// ISO-8601 timestamp of emission.
+    pub at: String,
+}
