+++ crates/pylon/src/handlers/knowledge/ingest.rs
@@ -2,9 +2,7 @@
 
 use axum::Json;
 use axum::extract::State;
-use serde::{Deserialize, Serialize};
 use tracing::instrument;
-use utoipa::ToSchema;
 
 use symbolon::types::Role;
 
@@ -12,39 +10,9 @@ use crate::error::{ApiError, BadRequestSnafu, ValidationFailedSnafu};
 use crate::extract::{Claims, require_role};
 use crate::state::KnowledgeState;
 
-/// Request body for knowledge ingestion.
-#[derive(Debug, Deserialize, ToSchema)]
-pub struct IngestRequest {
-    /// Raw content to ingest.
-    pub content: String,
-    /// Format: markdown, text, json, jsonl.
-    #[serde(default)]
-    pub format: String,
-    /// Nous agent ID that will own the extracted facts.
-    pub nous_id: String,
-}
-
-/// Per-fact error during ingestion.
-#[derive(Debug, Serialize, ToSchema)]
-pub struct IngestFactError {
-    /// Index of the fact in the batch.
-    pub index: usize,
-    /// Fact ID if available.
-    pub id: Option<String>,
-    /// Error message.
-    pub message: String,
-}
-
-/// Response for knowledge ingestion.
-#[derive(Debug, Serialize, ToSchema)]
-pub struct IngestResponse {
-    /// Number of facts successfully inserted.
-    pub inserted: usize,
-    /// Number of facts skipped due to errors.
-    pub skipped: usize,
-    /// Per-fact error details.
-    pub errors: Vec<IngestFactError>,
-}
+#[path = "ingest_dto.rs"]
+mod ingest_dto;
+pub use ingest_dto::{IngestFactError, IngestRequest, IngestResponse};
 
 /// POST /api/v1/knowledge/ingest
 ///
