+++ crates/pylon/src/handlers/knowledge/webhook.rs
@@ -2,9 +2,7 @@
 
 use axum::Json;
 use axum::extract::State;
-use serde::{Deserialize, Serialize};
 use tracing::instrument;
-use utoipa::ToSchema;
 
 use symbolon::types::Role;
 
@@ -12,31 +10,9 @@ use crate::error::{ApiError, BadRequestSnafu};
 use crate::extract::{Claims, require_role};
 use crate::state::KnowledgeState;
 
-/// Request body for webhook ingestion.
-///
-/// Accepts a single fact or a batch of facts from external systems
-/// (Slack, email, wiki, etc.).
-#[derive(Debug, Deserialize)]
-pub struct WebhookIngestRequest {
-    /// Nous agent ID that will own the facts.
-    pub nous_id: String,
-    /// Facts to insert.
-    pub facts: Vec<mneme::knowledge::Fact>,
-    /// Optional source identifier for provenance.
-    #[serde(default)]
-    pub source: Option<String>,
-}
-
-/// Response for webhook ingestion.
-#[derive(Debug, Serialize, ToSchema)]
-pub struct WebhookIngestResponse {
-    /// Number of facts successfully inserted.
-    pub inserted: usize,
-    /// Number of facts skipped due to errors.
-    pub skipped: usize,
-    /// Per-fact error details.
-    pub errors: Vec<crate::handlers::knowledge::ingest::IngestFactError>,
-}
+#[path = "webhook_dto.rs"]
+mod webhook_dto;
+pub use webhook_dto::{WebhookIngestRequest, WebhookIngestResponse};
 
 /// POST /api/v1/knowledge/ingest/webhook
 ///
