+++ crates/pylon/src/stream.rs
@@ -1,71 +1,8 @@
 //! SSE event types and hermeneus→SSE bridge.
 
-use serde::Serialize;
-use utoipa::ToSchema;
-
-/// SSE event emitted to the client during message streaming.
-#[derive(Debug, Clone, Serialize, ToSchema)]
-#[serde(tag = "type")]
-#[non_exhaustive]
-pub(crate) enum SseEvent {
-    /// Acknowledgment that the message was accepted and a turn is starting.
-    #[serde(rename = "message_start")]
-    MessageStart {
-        status: String,
-        /// Per-request correlation ID for distributed tracing across the pipeline.
-        #[serde(skip_serializing_if = "Option::is_none")]
-        request_id: Option<String>,
-    },
-
-    /// Incremental text output from the assistant.
-    #[serde(rename = "text_delta")]
-    TextDelta { text: String },
-
-    /// The assistant is invoking a tool.
-    #[serde(rename = "tool_use")]
-    ToolUse {
-        id: String,
-        name: String,
-        input: serde_json::Value,
-    },
-
-    /// Result returned from a tool execution.
-    #[serde(rename = "tool_result")]
-    ToolResult {
-        tool_use_id: String,
-        content: String,
-        is_error: bool,
-    },
-
-    /// Turn completed: final event in the stream.
-    #[serde(rename = "message_complete")]
-    MessageComplete {
-        stop_reason: String,
-        usage: UsageData,
-        /// Per-request correlation ID for distributed tracing across the pipeline.
-        #[serde(skip_serializing_if = "Option::is_none")]
-        request_id: Option<String>,
-    },
-
-    /// An error occurred during the turn.
-    #[serde(rename = "error")]
-    Error {
-        code: String,
-        message: String,
-        /// Per-request correlation ID for cross-system error tracing.
-        #[serde(skip_serializing_if = "Option::is_none")]
-        request_id: Option<String>,
-    },
-}
-
-/// Token usage summary sent with `message_complete`.
-#[derive(Debug, Clone, Serialize, ToSchema)]
-pub(crate) struct UsageData {
-    /// Tokens consumed by the system prompt and conversation history.
-    pub input_tokens: u64,
-    /// Tokens generated by the model in this turn.
-    pub output_tokens: u64,
-}
+#[path = "stream_dto.rs"]
+mod stream_dto;
+pub(crate) use stream_dto::{SseEvent, TurnOutcome, TurnStreamEvent, UsageData};
 
 impl SseEvent {
     /// SSE event name for the `event:` field.
@@ -80,73 +17,6 @@ impl SseEvent {
             Self::Error { .. } => "error",
         }
     }
-}
-
-/// SSE events for the turn streaming protocol (`POST /api/v1/sessions/stream`).
-///
-/// Used by `koilon` and the Signal integration. Unified with `SseEvent` naming:
-/// event types use the same `message_start`/`message_complete` vocabulary,
-/// and all fields use `snake_case` per API.md (#3271).
-#[derive(Debug, Clone, Serialize)]
-#[serde(tag = "type")]
-#[non_exhaustive]
-pub(crate) enum TurnStreamEvent {
-    /// Turn accepted — mirrors `SseEvent::MessageStart`.
-    #[serde(rename = "message_start")]
-    MessageStart {
-        session_id: String,
-        nous_id: String,
-        turn_id: String,
-        /// Per-request correlation ID for distributed tracing across the pipeline.
-        #[serde(skip_serializing_if = "Option::is_none")]
-        request_id: Option<String>,
-    },
-    /// Incremental extended-thinking output.
-    #[serde(rename = "thinking_delta")]
-    ThinkingDelta { text: String },
-    /// Incremental text output.
-    #[serde(rename = "text_delta")]
-    TextDelta { text: String },
-    /// Tool invocation started.
-    #[serde(rename = "tool_use")]
-    ToolUse {
-        tool_name: String,
-        tool_id: String,
-        input: serde_json::Value,
-    },
-    /// Tool execution is awaiting approval.
-    #[serde(rename = "tool_approval_required")]
-    ToolApprovalRequired {
-        turn_id: String,
-        tool_name: String,
-        tool_id: String,
-        input: serde_json::Value,
-        risk: String,
-        reason: String,
-    },
-    /// Tool approval decision resolved.
-    #[serde(rename = "tool_approval_resolved")]
-    ToolApprovalResolved { tool_id: String, decision: String },
-    /// Tool execution result.
-    #[serde(rename = "tool_result")]
-    ToolResult {
-        tool_name: String,
-        tool_id: String,
-        result: String,
-        is_error: bool,
-        duration_ms: u64,
-    },
-    /// Turn completed — mirrors `SseEvent::MessageComplete`.
-    #[serde(rename = "message_complete")]
-    MessageComplete { outcome: TurnOutcome },
-    /// An error occurred during the turn.
-    #[serde(rename = "error")]
-    Error {
-        message: String,
-        /// Per-request correlation ID for cross-system error tracing.
-        #[serde(skip_serializing_if = "Option::is_none")]
-        request_id: Option<String>,
-    },
 }
 
 impl TurnStreamEvent {
@@ -165,18 +35,4 @@ impl TurnStreamEvent {
             Self::Error { .. } => "error",
         }
     }
-}
-
-/// Turn completion data emitted in `MessageComplete` events.
-#[derive(Debug, Clone, Serialize)]
-pub(crate) struct TurnOutcome {
-    pub text: String,
-    pub nous_id: String,
-    pub session_id: String,
-    pub model: Option<String>,
-    pub tool_calls: usize,
-    pub input_tokens: u64,
-    pub output_tokens: u64,
-    pub cache_read_tokens: u64,
-    pub cache_write_tokens: u64,
 }
